+++ include/rands.h
@@ -27,9 +27,7 @@ typedef long RandListType;
 /* =================================================== */
 /*             Global Function Declarations            */
 /* --------------------------------------------------- */
-void RandSeed(
-    unsigned long initstate, unsigned long initseq, sw_random_t *pcg_rng
-);
+void RandSeed(size_t initstate, size_t initseq, sw_random_t *pcg_rng);
 
 double RandUni(sw_random_t *pcg_rng);
